requirements.txt
beautifulsoup4==4.8.1
config==0.3.9
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.17
django-bootstrap3==11.1.0
django-bootstrap4==1.0.1
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.8.0
gunicorn==19.9.0
Pillow==6.2.1
psycopg2==2.8.4
python-decouple==3.1
pytz==2019.3
soupsieve==1.9.4
whitenoise==4.1.4
